05_Automation/Scripts/requirements.txt
# BigSkyAg Automation Scripts Requirements
# Install with: pip install -r requirements.txt

# === CORE DEPENDENCIES ===
# These are required for basic functionality
pathlib2>=2.3.7; python_version < "3.4"  # For older Python versions

# === STORAGE PROVIDER DEPENDENCIES ===

# Google Drive API (BigSkyAg)
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Dropbox API (PaulysOps)
dropbox==11.36.2

# AWS S3 (Enterprise)
boto3==1.34.0
botocore==1.34.0

# === OPTIONAL DEPENDENCIES ===
# These provide enhanced functionality but aren't required

# Progress bars and enhanced output
tqdm==4.66.1

# Better logging and formatting
colorama==0.4.6

# Enhanced file operations
watchdog==3.0.0

# === SYSTEM DEPENDENCIES ===
# These are usually available on macOS but may need installation
# - zip (usually pre-installed)
# - rsync (usually pre-installed)

# === DEVELOPMENT DEPENDENCIES ===
# Only needed if developing or testing
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0

# === NOTES ===
# Core Python libraries (usually included with Python):
# - pathlib (Python 3.4+)
# - logging (Python standard library)
# - subprocess (Python standard library)
# - os (Python standard library)
# - shutil (Python standard library)
# - time (Python standard library)
# - typing (Python 3.5+)
# - abc (Python standard library)
# - json (Python standard library)
